pan course agg

+++ models/mechanical/coarse_aggregate.py
@@ -1294,7 +1294,7 @@ class CoarseAggregateMechanical(models.Model):
     sieve_visible = fields.Boolean("Sieve Analysis Visible",compute="_compute_visible")
 
     sieve_analysis_child_lines = fields.One2many('mechanical.coarse.aggregate.sieve.analysis.line','parent_id',string="Parameter")
-    total_sieve_analysis = fields.Integer(string="Total",compute="_compute_total_sieve")
+    total_sieve_analysis = fields.Float(string="Total",compute="_compute_total_sieve")
 
 
     def default_get(self, fields):
